// src/generated/client/worldmonitor/market/v1/service_client.ts
export interface MarketQuote {
  symbol: string;
  name: string;
  display: string;
  price: number;
  change: number;
  sparkline: number[];
}

export interface ListCryptoQuotesRequest {
  ids: string[];
}

export interface ListCryptoQuotesResponse {
  quotes: CryptoQuote[];
}

export interface CryptoQuote {
  name: string;
  symbol: string;
  price: number;
  change: number;
  sparkline: number[];
}

export interface ListCommodityQuotesRequest {
  symbols: string[];
}

export interface ListCommodityQuotesResponse {
  quotes: CommodityQuote[];
}

export interface CommodityQuote {
  symbol: string;
  name: string;
  display: string;
  price: number;
  change: number;
  sparkline: number[];
}

export interface GetSectorSummaryRequest {
  period: string;
}

export interface GetSectorSummaryResponse {
  sectors: SectorPerformance[];
}

export interface SectorPerformance {
  symbol: string;
  name: string;
  change: number;
}
// ...
export interface ListGulfQuotesResponse {
  quotes: GulfQuote[];
  rateLimited: boolean;
}

export interface GulfQuote {
  symbol: string;
  name: string;
  flag: string;
  country: string;
  type: string;
  price: number;
  change: number;
  sparkline: number[];
}

export interface GetMarketDashboardRequest {
}

export interface GetMarketDashboardResponse {
  stocks: MarketQuote[];
  commodities: CommodityQuote[];
  sectors: SectorPerformance[];
  crypto: CryptoQuote[];
  finnhubSkipped: boolean;
  skipReason: string;
  rateLimited: boolean;
}